MULTI-STEPPED CHANGES
"Change the HTML to adapt to a new CSS format, where the maximum width of the website will be 55em. Experiment with a new navbar that will have a "sticky" position and will move where the user goes. Redo the Projects and Professional Experience portion in the CSS, and make sure that for each second element in the Experience part, they alternate with each other."