"Remove unnecessary JS code + add CSS code to affix the menu button
I was experimenting with some JS code to fix the menu as the user was scrolling.Turned out it was totally unnecessary and so that led to its deletion.

I did not like the menu button being pushed against the top-right of the screen.In this case, I moved it slightly down and to the left. This was in the <dropbtn> class."